+++ CHANGELOG.md
@@ -87,3 +87,4 @@ comparable
 * Fixed unchecked cast warnings
 * Added initial support for Java 8 Streams
 * Replaced NodeMatcher with Java 8 Predicate
+* Expose the comparator of a sorting tree